+++ 30414/Hitter V. Andras/L04/trees/trees.cbp
@@ -31,6 +31,10 @@
 		<Compiler>
 			<Add option="-Wall" />
 		</Compiler>
+		<Unit filename="Listfunctions.h" />
+		<Unit filename="Treefunctions.h" />
+		<Unit filename="conversions.h" />
+		<Unit filename="def.h" />
 		<Unit filename="main.c">
 			<Option compilerVar="CC" />
 		</Unit>
